requirements.txt
asgiref==3.10.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-crontab==0.7.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.0.0
PyJWT==2.10.1
python-dateutil==2.8.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.6.3
psycopg2-binary==2.9.9